Change x,y types in constructor to match changes to entity class

+++ src/main/java/it/unibo/dragondash/model/impl/game/Character.java
@@ -11,7 +11,7 @@ public class Character extends Entity{
     int gravity;
     int maxHeight;
 
-    public Character(float x, float y, int width, int height) {
+    public Character(int x, int y, int width, int height) {
         super(x, y, width, height);
         
         setDefaultValues();
